# internal/handlers/register_user.go
package handlers

import (
	"encoding/json"
	"errors"
	"net/http"

	"github.com/jackc/pgx/v5/pgconn"

	"github.com/tamim1dev/task-manager/internal/database"
	"github.com/tamim1dev/task-manager/internal/models"
	"github.com/tamim1dev/task-manager/internal/utils"
)

// RegisterUser godoc
// @Summary Register a new user
// @Description Creates a new user account
// @Tags auth
// @Accept json
// @Produce json
// @Param request body models.CreateUserRequest true "Register payload"
// @Success 201 {object} models.User
// @Failure 400 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /auth/register [post]
func RegisterUser(w http.ResponseWriter, r *http.Request) {
	var newUser models.CreateUserRequest
	var returnUser models.ReturnNewUser

	jsonerr := json.NewDecoder(r.Body).Decode(&newUser)
	if jsonerr != nil {
		utils.ReturnError(w, http.StatusBadRequest, "Invalid request payload")
		return
	}
	defer r.Body.Close()

	// input validation
	if newUser.Name == "" || newUser.Email == "" || newUser.Password == "" {
		utils.ReturnError(w, http.StatusBadRequest, "Name, email, and password are required")
		return
	}

	password_hash, hasherr := utils.HashPassword(newUser.Password)
	if hasherr != nil {
		utils.ReturnError(w, http.StatusInternalServerError, "Error hashing password")
		return
	}

	query := `INSERT INTO users (name, email, password_hash) VALUES ($1, $2, $3) RETURNING name`
	user_err := database.DB.Pool.QueryRow(r.Context(), query, newUser.Name, newUser.Email, password_hash).Scan(&returnUser.Name)
	if user_err != nil {
		var pgErr *pgconn.PgError
		if errors.As(user_err, &pgErr) && pgErr.Code == "23505" {
			utils.ReturnError(w, http.StatusConflict, "Email already exists")
			return
		}
		utils.ReturnError(w, http.StatusInternalServerError, "Error creating user")
		return
	}
	utils.ReturnJson(w, http.StatusCreated, returnUser)
}